fix event index

+++ x/crosschain/client/cli/tx_vote_inbound.go
@@ -101,7 +101,7 @@ func CmdVoteInbound() *cobra.Command {
 				250_000,
 				argsCoinType,
 				argsAsset,
-				uint(argsEventIndex),
+				argsEventIndex,
 				protocolContractVersion,
 				isArbitraryCall,
 				argsInboundStatus,
